Create The Epitome of a Project
The Epitome of a Project means that a person will have fully prepared for a project and know all the requirements. Then the person will research, read, and study while articulating the contents of the project to assure that its viable and useful to the reader. The contents will be comprehensible and free of errors. A sufficient spell check (Grammarly) will be used for corrections and then professionalism.